refactor to use localStorage helpers
