+++ src/lib/sports/buncheses/BunchesesSnake.svelte
@@ -30,6 +30,7 @@
 	import {setCurrentTickDuration} from '$lib/SnakeGame';
 	import GameAudio from '$lib/GameAudio.svelte';
 	import Dimensions from '$lib/Dimensions.svelte';
+	import {assertNumber, getFromStorage, setInStorage} from '$lib/localStorage';
 
 	const storageKey = 'buncheses_high_score';
 	const clock = setClock(createClock({running: true}));
@@ -65,7 +66,7 @@
 
 	let applesEaten = 0;
 	let bunchesEaten = 0;
-	const highestClustersEaten = writable<number>(Number(localStorage.getItem(storageKey)) || 0);
+	const highestClustersEaten = writable(getFromStorage(storageKey, assertNumber) ?? 0);
 
 	const restart = (): void => {
 		if (!game) return;
@@ -87,7 +88,7 @@
 	// TODO is there a better place to do this? imperatively after updating the state?
 	$: if (bunchesEaten > $highestClustersEaten) {
 		$highestClustersEaten = bunchesEaten;
-		localStorage.setItem(storageKey, bunchesEaten + ''); // TODO use helper on store instead
+		setInStorage(storageKey, bunchesEaten);
 	}
 
 	const tick = (): boolean => {
